+++ packages/cli-kit/src/public/node/presets.test.ts
@@ -1,7 +1,7 @@
 import * as presets from './presets.js'
-import {inTemporaryDirectory, mkdir, write as fileWrite} from './file.js'
-import {join as pathJoin} from './path.js'
-import {encode as tomlEncode} from './toml.js'
+import {inTemporaryDirectory, mkdir, write as fileWrite} from '../../file.js'
+import {join as pathJoin} from '../../path.js'
+import {encode as tomlEncode} from '../../toml.js'
 import {describe, expect, test} from 'vitest'
 
 const preset1 = {
